DDCE-4769: Remove options at ReviewGoods page for exceeded allowances (#513)
* DDCE-4769: Remove options at ReviewGoods page for exceeded allowances

* DDCE-4769: Update dependencies and fix warnings

* DDCE-4769: Platform Mongo 5.0 TTL index change

+++ test/uk/gov/hmrc/merchandiseinbaggage/repositories/DeclarationJourneyRepositorySpec.scala
@@ -62,7 +62,7 @@ class DeclarationJourneyRepositorySpec extends BaseSpecWithApplication with Core
       val indices = declarationJourneyRepository.collection.listIndexes().toFuture().futureValue
 
       val actualTimeToLive =
-        indices.find(_.toString.contains("timeToLive")).get("expireAfterSeconds").asInt64().intValue()
+        indices.find(_.toString.contains("timeToLive")).get("expireAfterSeconds").asInt32().intValue()
 
       actualTimeToLive mustBe expectedTimeToLive
     }
